# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SubscriberInput {
  email: string;
}

export interface Subscriber {
  id: number;
  email: string;
  createdAt: string;
}

export interface ErrorResponse {
  error: string;
}

export interface AdminLoginInput {
  password: string;
}

export interface AdminStatus {
  authenticated: boolean;
}

export interface GuideStatus {
  id: string;
  title: string;
  hasFile: boolean;
  /** @nullable */
  fileName?: string | null;
}