# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStatus"
        "401":
          description: Wrong password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/subscribers:
    get:
      operationId: listSubscribers
      tags: [admin]
      summary: List all subscribers (admin only)
      responses:
        "200":
          description: Subscriber list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Subscriber"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    Subscriber:
      type: object
      required:
        - id
        - email
        - createdAt
      properties:
        id:
          type: integer
        email:
          type: string
        createdAt:
          type: string
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
    AdminLoginInput:
      type: object
      required:
        - password
      properties:
        password:
          type: string
    AdminStatus:
      type: object
      required:
        - authenticated
      properties:
        authenticated:
          type: boolean
...